+++ backend/hesco/src/main/java/com/lmkr/hesco/user/repository/RoleRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Short> {
+public interface RoleRepository extends JpaRepository<Role, Long> {
     Optional<Role> findByCode(String code);
 }
